requirements.txt
bertopic[visualization]
sentence-transformers
umap-learn
hdbscan
pandas
pyarrow
tqdm
emoji
regex
scikit-learn
python-dotenv
